Rename watch icons to correspond with AB files
When generating icons and splash screens the files that are generated do not correspond with the files used from plugin.
When people start using the plugin they will expect the generated images to work properly. If we do not have same filenames, they will not.
We need to change the filenames in the plugin to avoid breaking user experience – if someone has already generated their icons, they will have two empty if we change the filenames in AB.
So it is better from user point of view to change them in the plugin.